refactor(CodeGeneration,Commands): abstracted cook related logic from CookCommand and plopFile
