+++ src/cli.ts
@@ -1,5 +1,5 @@
-import { CookCommand } from "@commands/CookCommand";
 import { HelloCommand } from "@commands/HelloCommand";
+import { CookCommandCommand } from "@commands/cook/CookCommandCommand";
 import { Command } from "commander";
 
 const program = new Command();
@@ -7,6 +7,6 @@ const program = new Command();
 program.name("CindroX-cli").description("CLI tool for the CindroX framework").version("1.0.0");
 
 program.addCommand(new HelloCommand());
-program.addCommand(new CookCommand());
+program.addCommand(new CookCommandCommand());
 
 program.parse();
